Move `libeeprom` to `firmware-base`
