+++ lib/firmware-base/eeprom.h
@@ -2,6 +2,7 @@
 #define EEPROM_H
 
 #include <esp_err.h>
+
 #include <stddef.h>
 #include <stdint.h>
 
@@ -11,4 +12,4 @@ esp_err_t eeprom_read(uint16_t addr, uint8_t *data, size_t len);
 esp_err_t eeprom_write(uint16_t addr, const uint8_t *data, size_t len);
 esp_err_t eeprom_write_byte(uint16_t addr, uint8_t data);
 
-#endif
+#endif	// EEPROM_H
