+++ lib/firmware-base/eeprom_ember.c
@@ -1,0 +1,10 @@
+#ifdef _EMBER
+
+#include "eeprom.h"
+#include <ember_taskglue.h>
+
+ember_rate_funcs_S eeprom_rf = {
+    .call_init = eeprom_init,
+};
+
+#endif	// _EMBER
